Working on bonus

+++ srcs/parsing_2.c
@@ -6,7 +6,7 @@
 /*   By: tgernez <tgernez@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/07 14:20:31 by tgernez           #+#    #+#             */
-/*   Updated: 2022/12/08 11:08:32 by tgernez          ###   ########.fr       */
+/*   Updated: 2022/12/08 13:50:18 by tgernez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,7 +98,7 @@ void	iso(int ***pts, int *dims, double ang)
 	}
 }
 
-void	alt_adding(int ***pts, int *dims, t_point **map, int scale)
+void	alt_adding(int ***pts, int *dims, t_point **map, double scale)
 {
 	int	i;
 	int	j;
